+++ lib-dmx/include/h3/dmx.h
@@ -42,7 +42,7 @@ public:
 	// RDM
 	void RdmSendRaw(uint32_t nPort, const uint8_t *pRdmData, uint32_t nLength) override;
 	const uint8_t *RdmReceive(uint32_t nPort) override;
-	const uint8_t *RdmReceiveTimeOut(uint32_t nPort, uint32_t nTimeOut) override;
+	const uint8_t *RdmReceiveTimeOut(uint32_t nPort, uint16_t nTimeOut) override;
 	uint32_t RdmGetDateReceivedEnd() override;
 
 	// DMX
